Move TinyJson to its own folder
