+++ TechTalk.SpecFlow/TinyJson/JSONWriter.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Runtime.Serialization;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace TechTalk.SpecFlow
+namespace TechTalk.SpecFlow.TinyJson
 {
     public class JsonSerializerSettings
     {
@@ -22,6 +20,7 @@ namespace TechTalk.SpecFlow
         }
     }
 
+    // https://github.com/zanders3/json
     //Really simple JSON writer
     //- Outputs JSON structures from an object
     //- Really simple API (new List<int> { 1, 2, 3 }).ToJson() == "[1,2,3]"
